Fix requested issues
