+++ app.js
@@ -3,7 +3,7 @@ const express = require("express");
 const path = require("path");
 const cors = require("cors");
 const cron = require("node-cron");
-const price = require("./server/controllers/price");
+const price = require("./server/services/price");
 const app = express();
 
 app.use(function(req, res, next) {
@@ -25,7 +25,7 @@ app.use(cors());
 
 const server = app.listen(process.env.PORT || 3000);
 
-var task = cron.schedule("59 23 * * *", price.checkForAllWatched, true);
+var task = cron.schedule("59 23 * * *", price.checkAllWatched, true);
 
 task.start();
